refactor: 패키지명 도메인명으로 수정
