+++ src/main/kotlin/calculator/domain/InputParser.kt
@@ -1,4 +1,4 @@
-package step1.domain
+package calculator.domain
 
 import java.util.LinkedList
