+++ src/main/kotlin/calculator/domain/Operator.kt
@@ -1,4 +1,4 @@
-package step1.domain
+package calculator.domain
 
 import java.lang.IllegalArgumentException
