+++ src/test/kotlin/calculator/domain/ParsedExpressionTest.kt
@@ -1,4 +1,4 @@
-package step1.domain
+package calculator.domain
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
